Changed the way PageAllocator handles decreasing capacity

+++ DX12Engine/GpuHeapLocation.h
@@ -2,9 +2,9 @@
 
 struct GpuHeapLocation
 {
-	unsigned int heapIndex;
-	unsigned int heapOffsetInPages;
+	unsigned long heapIndex;
+	unsigned short heapOffsetInPages;
 
 	GpuHeapLocation() {} //so containers don't zero heapIndex and heapOffsetInPages when they doen't need to
-	GpuHeapLocation(unsigned int heapIndex, unsigned int heapOffsetInPages) : heapIndex(heapIndex), heapOffsetInPages(heapOffsetInPages) {} //needed because of other constructor
+	GpuHeapLocation(unsigned long heapIndex, unsigned short heapOffsetInPages) : heapIndex(heapIndex), heapOffsetInPages(heapOffsetInPages) {} //needed because of other constructor
 };
